# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Panadería Polvillo - API REST
  description: |
    API REST para la tienda online y panel de gestión de **Panadería Polvillo**.
    
    ## 🔐 Autenticación
    
    Esta API usa **JWT Bearer Token** para proteger los endpoints.
    
    ### Pasos para autenticarse:
    1. Usa `POST /v1/auth/register` para crear una cuenta o usa las credenciales de prueba
    2. Usa `POST /v1/auth/login` con tus credenciales
    3. Copia el **access_token** de la respuesta
    4. Haz clic en el botón **"Authorize"** 🔓 arriba
    5. Pega el token (sin "Bearer ", solo el token)
    6. Haz clic en **"Authorize"** y luego **"Close"**
    
    ### Credenciales de prueba:
    
    | Rol | Email | Password |
    |-----|-------|----------|
    | Admin | admin@panaderia.com | admin123 |
    | Cliente | cliente@test.com | test123 |
    
    ## 📁 Categorías disponibles
    - Despensa y básicos
    - Conservas y Enlatados
    - Aceites, Vinagres y Salsas
    - Bebidas y Bodega
    - Charcutería
    - Dulces
    - Panadería
    
  version: "1.0.0"
  contact:
    name: Francisco Mercado
    email: admin@panaderia.com
servers:
  - url: http://localhost:3001/v1
    description: Servidor de desarrollo local
  - url: https://franciscomercado-pi-back.onrender.com/v1
    description: Servidor de producción (Render)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint POST /v1/auth/login
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Descripción del error
...
    Producto:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        nombre:
          type: string
          example: Barra Campesina
        descripcion:
          type: string
          example: Pan tradicional de corteza crujiente
        precio:
          type: number
          format: float
          example: 1.20
        imagen_url:
          type: string
          example: https://res.cloudinary.com/...
        categoria:
          type: string
          enum: 
            - Despensa y básicos
            - Conservas y Enlatados
            - Aceites, Vinagres y Salsas
            - Bebidas y Bodega
            - Charcutería
            - Dulces
            - Panadería
        stock:
          type: integer
          example: 50
        stock_minimo:
          type: integer
          example: 20
        stock_bajo:
          type: boolean
          example: false
        estado_inventario:
          type: string
          enum: [disponible, bajo, agotado]
        disponible:
          type: boolean
          example: true
        peso:
          type: integer
          description: Peso en gramos
          example: 250
        ingredientes:
          type: string
        alergenos:
          type: array
          items:
            type: string
          example: ["gluten"]
        destacado:
          type: boolean
          example: true
...
    CarritoItem:
      type: object
      properties:
        _id:
          type: string
        producto:
          $ref: '#/components/schemas/Producto'
        cantidad:
          type: integer
          minimum: 1
        precio_unitario:
          type: number
    Carrito:
      type: object
      properties:
        _id:
          type: string
        usuario:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CarritoItem'
        total:
          type: number
...
  responses:
    UnauthorizedError:
      description: Token no proporcionado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token de autenticación inválido o expirado
...
paths:
...
  /carrito:
    get:
      tags: [Carrito]
      summary: Obtener carrito del usuario
      operationId: getCarrito
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Carrito del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Carrito'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags: [Carrito]
      summary: Vaciar carrito
      operationId: clearCarrito
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Carrito vaciado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Carrito'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...